pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.3",
    "piexif>=1.1.3",
    "pillow>=12.1.1",
]
